python/sliced_bread/requirements.txt
asgiref==3.3.4
Django==3.2.1
pytz==2021.1
sqlparse==0.4.1
django-shortuuidfield==0.1.3
django-cors-headers==3.10.0
djangorestframework==3.12.4
